LPS-74933 Move util methods
