+++ modules/sdk/gradle-plugins-defaults/src/main/java/com/liferay/gradle/plugins/defaults/internal/LiferayRelengPlugin.java
@@ -27,6 +27,7 @@ import com.liferay.gradle.plugins.defaults.LiferayOSGiDefaultsPlugin;
 import com.liferay.gradle.plugins.defaults.LiferayThemeDefaultsPlugin;
 import com.liferay.gradle.plugins.defaults.internal.util.FileUtil;
 import com.liferay.gradle.plugins.defaults.internal.util.GitUtil;
+import com.liferay.gradle.plugins.defaults.internal.util.GradlePluginsDefaultsUtil;
 import com.liferay.gradle.plugins.defaults.internal.util.GradleUtil;
 import com.liferay.gradle.plugins.defaults.tasks.MergeFilesTask;
 import com.liferay.gradle.plugins.defaults.tasks.ReplaceRegexTask;
@@ -307,7 +308,7 @@ public class LiferayRelengPlugin implements Plugin<Project> {
 				public boolean isSatisfiedBy(Task task) {
 					Project project = task.getProject();
 
-					if (!GradleUtil.isTestProject(project) &&
+					if (!GradlePluginsDefaultsUtil.isTestProject(project) &&
 						_hasProjectDependencies(project)) {
 
 						return true;
@@ -648,7 +649,7 @@ public class LiferayRelengPlugin implements Plugin<Project> {
 
 					if (GradleUtil.hasStartParameterTask(
 							project, task.getName()) ||
-						!GradleUtil.isSnapshot(project)) {
+						!GradlePluginsDefaultsUtil.isSnapshot(project)) {
 
 						return true;
 					}
@@ -710,7 +711,7 @@ public class LiferayRelengPlugin implements Plugin<Project> {
 	}
 
 	private void _configureTaskPrintStaleArtifactForOSGi(Task task) {
-		if (GradleUtil.isTestProject(task.getProject())) {
+		if (GradlePluginsDefaultsUtil.isTestProject(task.getProject())) {
 			task.setEnabled(false);
 		}
 	}
@@ -756,7 +757,7 @@ public class LiferayRelengPlugin implements Plugin<Project> {
 
 		Project project = writeArtifactPublishCommandsTask.getProject();
 
-		if (GradleUtil.isTestProject(project)) {
+		if (GradlePluginsDefaultsUtil.isTestProject(project)) {
 			writeArtifactPublishCommandsTask.setEnabled(false);
 		}
