+++ modules/sdk/gradle-plugins-defaults/src/main/java/com/liferay/gradle/plugins/defaults/internal/MavenDefaultsPlugin.java
@@ -17,6 +17,7 @@ package com.liferay.gradle.plugins.defaults.internal;
 import com.liferay.gradle.plugins.BaseDefaultsPlugin;
 import com.liferay.gradle.plugins.defaults.LiferayOSGiDefaultsPlugin;
 import com.liferay.gradle.plugins.defaults.internal.util.FileUtil;
+import com.liferay.gradle.plugins.defaults.internal.util.GradlePluginsDefaultsUtil;
 import com.liferay.gradle.plugins.defaults.internal.util.GradleUtil;
 
 import java.io.File;
@@ -66,7 +67,7 @@ public class MavenDefaultsPlugin extends BaseDefaultsPlugin<MavenPlugin> {
 					return;
 				}
 
-				if (GradleUtil.isSnapshot(project)) {
+				if (GradlePluginsDefaultsUtil.isSnapshot(project)) {
 					File relengDir = new File(portalRootDir, "modules/.releng");
 
 					if (relengDir.exists()) {
