+++ modules/sdk/gradle-plugins-defaults/src/main/java/com/liferay/gradle/plugins/defaults/internal/util/GradlePluginsDefaultsUtil.java
@@ -40,6 +40,10 @@ public class GradlePluginsDefaultsUtil {
 	public static final String[] JSON_VERSION_FILE_NAMES =
 		{"npm-shrinkwrap.json", "package-lock.json", "package.json"};
 
+	public static final String SNAPSHOT_PROPERTY_NAME = "snapshot";
+
+	public static final String SNAPSHOT_VERSION_SUFFIX = "-SNAPSHOT";
+
 	public static final String TMP_MAVEN_REPOSITORY_DIR_NAME = ".m2-tmp";
 
 	public static final Pattern jsonVersionPattern = Pattern.compile(
@@ -125,6 +129,71 @@ public class GradlePluginsDefaultsUtil {
 					}
 
 				});
+		}
+	}
+
+	public static boolean isPrivateProject(Project project) {
+		String path = project.getPath();
+
+		if (path.startsWith(":private:")) {
+			return true;
+		}
+
+		return false;
+	}
+
+	public static boolean isSnapshot(Project project) {
+		String version = String.valueOf(project.getVersion());
+
+		if (version.endsWith(SNAPSHOT_VERSION_SUFFIX)) {
+			return true;
+		}
+
+		return false;
+	}
+
+	public static boolean isSnapshot(Project project, String... propertyNames) {
+		boolean snapshot = false;
+
+		if (project.hasProperty(SNAPSHOT_PROPERTY_NAME)) {
+			snapshot = GradleUtil.getProperty(
+				project, SNAPSHOT_PROPERTY_NAME, true);
+		}
+
+		if (!snapshot) {
+			for (String propertyName : propertyNames) {
+				if (project.hasProperty(propertyName) &&
+					GradleUtil.getProperty(project, propertyName, true)) {
+
+					snapshot = true;
+
+					break;
+				}
+			}
+		}
+
+		return snapshot;
+	}
+
+	public static boolean isTestProject(Project project) {
+		String projectName = project.getName();
+
+		if (projectName.endsWith("-test")) {
+			return true;
+		}
+
+		return false;
+	}
+
+	public static void setProjectSnapshotVersion(
+		Project project, String... propertyNames) {
+
+		String version = String.valueOf(project.getVersion());
+
+		if (isSnapshot(project, propertyNames) &&
+			!version.endsWith(SNAPSHOT_VERSION_SUFFIX)) {
+
+			project.setVersion(version + SNAPSHOT_VERSION_SUFFIX);
 		}
 	}
